Add files via upload
This is a simple program to read an array of Student Objects, sort them based on the rollNo(attribute in the Student class). For brevity, I just considered the array of size 4. Here, sorting is done manually without using Comparators.
The Student is another class with basic details like name, rollNo, marks.