Count ways to reach the n'th stair
 There are n stairs, a person standing at the bottom wants to reach the top. The person can climb either 1 stair or 2 stairs at a time. 
This program counts the number of ways, the person can reach the top.

Here we are not solving it using traditional recursion process, but dynamic programming is being used to avoid the drawbacks in recursion process( re computation of the already computed values ). This is done by maintaining an array of all the previously computed values.